DSCEngine

+++ src/DecentralizedStableCoin.sol
@@ -23,7 +23,6 @@
 // private
 // view & pure functions
 
-
 pragma solidity ^0.8.18;
 
 import {ERC20Burnable, ERC20} from "@openzeppelin/contracts/token/ERC20/extensions/ERC20Burnable.sol";
@@ -34,22 +33,19 @@ import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
  * Collateral: Exogenous (ETH & BTC)
  * Minting: Algorithmic
  * Relatice Stability: Pegged to USD
- * 
+ *
  * This is the contrat meant to be governed by DSCEngine. This contract is hust the ERC20 implementation of our stablecoin system
  */
 
-contract DecentralizedStableCoin is ERC20Burnable, Ownable{
-
+contract DecentralizedStableCoin is ERC20Burnable, Ownable {
     error DecentralizedStableCoin__BurnAmountExceedsBalnce();
     error DecentralizedStableCoin__NotZeroAddress();
     error DecentralizedStableCoin__MustBeMoreThanZero();
 
-    constructor() ERC20("DecentralizedStableCoin", "DSC") {
-        
-    }
+    constructor() ERC20("DecentralizedStableCoin", "DSC") {}
 
-    function burn(uint _amount) public override onlyOwner {
-        uint balance = balanceOf(msg.sender);
+    function burn(uint256 _amount) public override onlyOwner {
+        uint256 balance = balanceOf(msg.sender);
         if (_amount <= 0) {
             revert DecentralizedStableCoin__MustBeMoreThanZero();
         }
@@ -59,7 +55,7 @@ contract DecentralizedStableCoin is ERC20Burnable, Ownable{
         super.burn(_amount);
     }
 
-    function mint(address _to, uint256 _amount) external onlyOwner returns(bool) {
+    function mint(address _to, uint256 _amount) external onlyOwner returns (bool) {
         if (_to == address(0)) {
             revert DecentralizedStableCoin__NotZeroAddress();
         }
